+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### 0.5.0 (2020-03-22)
+1. **Simpletools\Db\Cassandra\TempTable**
+    1. Added register functions to improve cleanup and minimise tables leaking 
+    2. Added `::registerAutoShutdown()` to enable on terminate process signal and shutdown auto cleanup 
+    3. Added `::registerMaxSize()` to specify max size of opened TempTables in the current process
+    4. Added `::cleanup($signal=0)` to start register cleanup and temp tables deletion, otherwise automatically by with `::registerShutdown()`
+    5. Added `->drop()` to enable individual object table cleanup
+    6. Added `::listActiveTempTables($keyspace)` to get list of all active temp tables in a given keyspace
+
 ### 0.4.0 (2020-02-18)
 1. **Simpletools\Db\Cassandra**
     1. Added integration with `Simpletools\Db\Replicator` to replicate data between databases